requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
colorama==0.4.6
Faker==37.3.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
idna==3.10
iniconfig==2.1.0
kafka-python==2.2.10
packaging==25.0
pluggy==1.6.0
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
prometheus_client
prometheus_fastapi_instrumentator
jsonschema

